+++ apps/server/src/modules/learnroom/controller/dashboard.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { DashboardEntity } from '@shared/domain';
+import { DashboardEntity, EntityId, GridPosition, GridElement, DefaultGridReference } from '@shared/domain';
 import { DashboardUc } from '../uc/dashboard.uc';
 import { DashboardController } from './dashboard.controller';
 import { DashboardResponse } from './dto';
@@ -19,6 +19,9 @@ describe('dashboard uc', () => {
 						getUsersDashboard() {
 							throw new Error('Please write a mock for DashboardRepo.getUsersDashboard.');
 						},
+						moveElementOnDashboard(dashboardId: EntityId, from: GridPosition, to: GridPosition) {
+							throw new Error('Please write a mock for DashboardRepo.getUsersDashboard.');
+						},
 					},
 				},
 			],
@@ -36,6 +39,44 @@ describe('dashboard uc', () => {
 			});
 			const response = await controller.findForUser();
 
+			expect(response instanceof DashboardResponse).toEqual(true);
+		});
+	});
+
+	describe('moveElement', () => {
+		it('should call uc', async () => {
+			const spy = jest
+				.spyOn(uc, 'moveElementOnDashboard')
+				.mockImplementation((dashboardId: EntityId, from: GridPosition, to: GridPosition) => {
+					const dashboard = new DashboardEntity(dashboardId, {
+						grid: [
+							{
+								pos: to,
+								gridElement: new GridElement('elementId', new DefaultGridReference('referenceId', 'Mathe')),
+							},
+						],
+					});
+					return Promise.resolve(dashboard);
+				});
+			await controller.moveElement('dashboardId', { from: { x: 1, y: 2 }, to: { x: 2, y: 1 } });
+			expect(spy).toHaveBeenCalledWith('dashboardId', { x: 1, y: 2 }, { x: 2, y: 1 });
+		});
+
+		it('should return a dashboard', async () => {
+			jest
+				.spyOn(uc, 'moveElementOnDashboard')
+				.mockImplementation((dashboardId: EntityId, from: GridPosition, to: GridPosition) => {
+					const dashboard = new DashboardEntity(dashboardId, {
+						grid: [
+							{
+								pos: to,
+								gridElement: new GridElement('elementId', new DefaultGridReference('referenceId', 'Mathe')),
+							},
+						],
+					});
+					return Promise.resolve(dashboard);
+				});
+			const response = await controller.moveElement('dashboardId', { from: { x: 1, y: 2 }, to: { x: 2, y: 1 } });
 			expect(response instanceof DashboardResponse).toEqual(true);
 		});
 	});