+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ### Added
 
 - BC-121 - add console script to delete soft-deleted files
-- BC-343 - run unstable e2e tests on label `run unstable tests`
+- BC-319 - add route to change the position of a course dashboard element
+- BC-343 - run unstable e2e tests on label `run unstable tests` 
 
 ### Changed